requirements.txt
numpy
joblib
scikit-learn

#rafdb
onnxruntime>=1.15.1
facenet-pytorch>=2.5.2  
opencv-python
torch               
Pillow
matplotlib